Identify unique taxid combinations and run NCBI topology on subset
Reduces the run time of sam2lca by identifying all unique combinations
of taxids that were observed across all reads and by running the
expensive NCBI topology function of ete3 on these. For samples that have
a lot of aligned reads but only a few combinations of taxids, this will
decrease the number of lookups strongly.