Escape case when no reads pass the minimal percent identity threshold
When no reads pass the minimal percent identity threshold to the
reference genome, the object taxid_info_dict is empty. In this case,
converting it to a pandas DataFrame object results into an empty
DataFrame that does not have any columns. To avoid throwing a KeyError
trying to access some columns by name, the test is added prior to it.
If the dictionary is empty, no output table will be written and no BAM
files will be produced but just the empty JSON file will be written to
file.